Move from definitions directory for cleanup
